% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CreateChecklist.R
\name{CreateChecklist}
\alias{CreateChecklist}
\title{CreateChecklist}
\usage{
CreateChecklist(specieslist, export = "export")
}
\arguments{
\item{specieslist}{path to list of species}

\item{export}{name of your exported checklist file}
}
\value{
txt file
}
\description{
Create a custom checklist with species shortnames
}
\examples{
## NOT RUN
if (interactive()) {
CreateChecklist(specieslist = paste0(path.package("Rveg"),
"/extdata/SpeciesList"))
}

}
