% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Rveg-package.R
\docType{package}
\name{Rveg-package}
\alias{Rveg}
\alias{Rveg-package}
\title{Rveg: Digitazion of phytosociological releves}
\description{
Simple and fast tool to transforming phytosociological vegetation data into the digital form for following analysis.
}
\author{
\strong{Maintainer}: Premysl Kral \email{kralp@fzp.czu.cz} (\href{https://orcid.org/0000-0002-2887-651x}{ORCID})

Authors:
\itemize{
  \item Jan Douda \email{douda@fzp.czu.cz} (\href{https://orcid.org/0000-0002-1205-364X}{ORCID})
}

}
\keyword{internal}
