% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RvegCheck.R
\name{RvegCheck}
\alias{RvegCheck}
\title{Checking your DATABASE for duplicity}
\usage{
RvegCheck(DATABASE)
}
\arguments{
\item{DATABASE}{name of csv files for releve table and header}

\item{SAVE}{name of exporting database}

\item{checklist}{custom checklist}

\item{extrahead}{extra rows in header}
}
\value{
Export two csv files, one for releve and one for header
}
\description{
Checking your DATABASE for duplicity
}
\examples{

modify(SAVE = "test_database")

}
