% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RvegCombine.R
\name{RvegCombine}
\alias{RvegCombine}
\title{Combining species in the database}
\usage{
RvegCombine(database, export)
}
\arguments{
\item{database}{name of loading database}

\item{export}{name of the export database}

\item{save}{name of exported databes withou extension}

\item{head}{logical value if you want to export header data aswell}
}
\value{
Rveg database containing two csv files, header data and releve data.
}
\description{
Combining species in the database
}
\examples{
RvegCombine(x, y)



}
