% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RvegMERGE.r
\name{RvegMERGE}
\alias{RvegMERGE}
\title{Merge two Rveg databases}
\usage{
RvegMERGE(x, y, save = "export_merge", head = T)
}
\arguments{
\item{x}{name of first database without extension}

\item{y}{name of second database without extension}

\item{save}{name of exported databes withou extension}

\item{head}{logical value if you want to export header data aswell}
}
\value{
Rveg database containing two csv files, header data and releve data.
}
\description{
Merge two Rveg databases
}
\examples{
RvegMERGE(x, y)



}
