% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RvegToJuice.R
\name{RvegToTv}
\alias{RvegToTv}
\title{RvegToTv}
\usage{
RvegToTv(database, export = "export", ver = 3, checklist = "default")
}
\arguments{
\item{database}{path to Rveg database}

\item{export}{name of your exported Tv file}

\item{ver}{version of TURBOVEG}

\item{checklist}{checklist to match Fullnames}
}
\value{
csv file
}
\description{
Export Turboveg csv compatible file
}
\examples{
## NOT RUN
if (interactive()) {
  RvegToTv(database = paste0(
    path.package("Rveg"),
    "/extdata/example_db"
  ))
}

}
