% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RvegToJuice.R
\name{tvToRveg}
\alias{tvToRveg}
\title{Export database Juice software compatible format}
\usage{
tvToRveg(tv, export = "export", ...)
}
\arguments{
\item{Data}{name of your database without extension}
}
\value{
csv file
}
\description{
Export database Juice software compatible format
}
\examples{

tvToRveg(Database)

}
