% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addReleve.R
\name{addReleve}
\alias{addReleve}
\title{addReleve}
\usage{
addReleve(DATABASE = "NEW", SAVE, checklist = "default", extrahead = NULL)
}
\arguments{
\item{DATABASE}{name of csv files for releve table and header - database}

\item{SAVE}{name of exporting database}

\item{checklist}{custom checklist}

\item{extrahead}{extra rows in header}
}
\value{
export two csv files, one for releve and one for header
}
\description{
Writing and editing your releves
}
