% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addReleve.R
\name{addReleve}
\alias{addReleve}
\title{addReleve}
\usage{
addReleve(
  DATABASE = "NEW",
  SAVE = "default",
  checklist = "default",
  extrahead = NULL,
  start = TRUE
)
}
\arguments{
\item{DATABASE}{name of csv files for releve table and header - database}

\item{SAVE}{name of exporting database}

\item{checklist}{custom checklist}

\item{extrahead}{extra rows in header}

\item{start}{Boolean to start immediately digitizing first releve}
}
\value{
export two csv files, one for releve and one for header (Rveg database)
}
\description{
Digitizing and editing your releves - Rveg database
}
\examples{
## NOT RUN
if (interactive()) {
  addReleve()
}

}
