% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addReleve.R
\name{addRELEVE}
\alias{addRELEVE}
\title{Writing end editing your releves}
\usage{
addRELEVE(DATARELE = "NEW", DATAhead = "NEW", SAVE, checklist = "default")
}
\arguments{
\item{DATARELE}{csv file for releve table}

\item{DATAhead}{csv file for environmental variables}
}
\value{
Export two csv files
}
\description{
Writing end editing your releves
}
\examples{

addRELEVE(SAVE = EXAMPLE)

}
