% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CreateChecklist.R
\name{CreateChecklist}
\alias{CreateChecklist}
\title{CreateChecklist}
\usage{
CreateChecklist(specieslist, export = "export")
}
\arguments{
\item{specieslist}{path to list of species}

\item{export}{name of your exported checklist file}
}
\value{
txt file used as checklist in Rveg functions
}
\description{
Create a custom checklist with species ShortNames
}
\examples{
## NOT RUN
if (interactive()) {
# using txt file
CreateChecklist(specieslist = paste0(path.package("Rveg"),
"/extdata/SpeciesList"))

# creating from wcvp using rWCVP package
stopifnot(requireNamespace("rWCVP"))
a <- wcvp_checklist(area_codes = "QUE")
a <- unique(a$taxon_name)
CreateChecklist(a,export = "wcvp_que")
}

}
