% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Rveg-package.R
\docType{package}
\name{Rveg-package}
\alias{Rveg}
\alias{Rveg-package}
\title{Rveg: Digitization of Phytosociological Relevés}
\description{
Simple and fast tool for transforming phytosociological vegetation data into digital form for the following analysis.
For detailed guide run \code{vignette("Rveg","Rveg")}
or check the quick start guide on \url{https://github.com/sesitcsl2/Rveg}
}
\references{
Danihelka, Chrtek, and Kaplan (2012, ISSN:00327786), "Checklist of vascular plants of the czech republic. Preslia".

Hennekens, and Schaminée (2001) \url{doi:10.2307/3237010}, "TURBOVEG, a comprehensive data base management system for vegetation data. Journal of Vegetation Science".

Tichý (2002) \url{doi:10.1111/j.1654-1103.2002.tb02069.x}, "JUICE, software for vegetation classification. Journal of Vegetation Science".

Wickham, François, Henry, Müller (2022) \url{https://CRAN.R-project.org/package=dplyr}, "dplyr: A Grammar of Data Manipulation".
}
\seealso{
Useful links:
\itemize{
  \item \url{https://plant-ecology-lab-czu.com/rveg/}
}

}
\author{
\strong{Maintainer}: Přemysl Král \email{kralp@fzp.czu.cz} (\href{https://orcid.org/0000-0002-2887-651X}{ORCID})

Authors:
\itemize{
  \item Jan Douda \email{douda@fzp.czu.cz} (\href{https://orcid.org/0000-0002-1205-364X}{ORCID})
}

}
\keyword{internal}
