% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Rveg-package.R
\docType{package}
\name{Rveg-package}
\alias{Rveg}
\alias{Rveg-package}
\title{Rveg: Digitization of Phytosociological Releves}
\description{
Simple and fast tool for transforming phytosociological vegetation data into digital form for the following analysis.
For detailed guide run \code{vignette("Rveg","Rveg")}
or check the quick start guide on \url{https://github.com/sesitcsl2/Rveg}
}
\references{
DANIHELKA J., J. CHRTEK a Z. KAPLAN, 2012. Checklist of vascular plants of the czech republic. Preslia. 84(3), 647–811. ISSN 00327786.

HENNEKENS S.M. a J.H.J. SCHAMINÉE, 2001. TURBOVEG, a comprehensive data base management system for vegetation data. Journal of Vegetation Science. 12(4), 589–591. ISSN 1100-9233. Available at: \url{doi:10.2307/3237010}

TICHÝ L., 2002. JUICE, software for vegetation classification. Journal of Vegetation Science. 13(3), 451. ISSN 1100-9233. Available at: \url{doi:10.1658/1100-9233(2002)013[0451:jsfvc]2.0.co;2}

WICKHAM H., R. FRANCOIS, L HENRY, K MULLER, 2022. dplyr: A Grammar of Data Manipulation. R package version 1.0.10. Available at: \url{https://CRAN.R-project.org/package=dplyr}
}
\author{
\strong{Maintainer}: Premysl Kral \email{kralp@fzp.czu.cz} (\href{https://orcid.org/0000-0002-2887-651x}{ORCID})

Authors:
\itemize{
  \item Jan Douda \email{douda@fzp.czu.cz} (\href{https://orcid.org/0000-0002-1205-364X}{ORCID})
}

}
\keyword{internal}
