% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RvegCheck.R
\name{RvegCheck}
\alias{RvegCheck}
\title{RvegCheck}
\usage{
RvegCheck(
  DATABASE,
  fullnames = FALSE,
  export = "export",
  checklist = "default"
)
}
\arguments{
\item{DATABASE}{name of csv files for releve table and header - database}

\item{fullnames}{logical value if you want to add fullnames to the database}

\item{export}{name of exporting database}

\item{checklist}{checklist used to match shortnames with species name}
}
\value{
Export csv file releve table
}
\description{
Checking your DATABASE for duplicity and allowing to export table with full species name (not editable anymore).
}
