% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RvegCombine.R
\name{RvegCombine}
\alias{RvegCombine}
\title{RvegCombine}
\usage{
RvegCombine(database, export = "export", checklist = "default")
}
\arguments{
\item{database}{name of the loading database}

\item{export}{name of the exported database}

\item{checklist}{checklist to be used}
}
\value{
export two csv files, one for releve and one for header
}
\description{
Merging species or layers in the database
}
\examples{
## NOT RUN
if (interactive()) {
  RvegCombine(database = paste0(
    path.package("Rveg"),
    "/extdata/example_db"
  ))
}

}
