% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RvegLoad.R
\name{RvegLoad}
\alias{RvegLoad}
\title{RvegLoad}
\usage{
RvegLoad(DATABASE = "export", checklist = "default")
}
\arguments{
\item{DATABASE}{name of csv files for releve table and header - database}

\item{checklist}{used checklist}
}
\value{
read the database in one object
}
\description{
Writing and editing your releves
}
\examples{
## NOT RUN
if (interactive()) {
  RvegLoad
}

}
