% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RvegLoad.R
\name{RvegLoad}
\alias{RvegLoad}
\title{RvegLoad}
\usage{
RvegLoad(DATABASE = "default", CustomScale = FALSE, checklist = "default")
}
\arguments{
\item{DATABASE}{name of Rveg database}

\item{CustomScale}{logical values if different than predefined scale was used during the database creation}

\item{checklist}{used checklist}
}
\value{
read the database in one object
}
\description{
Reading your Rveg database
}
\examples{

  RvegLoad()

}
