% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RvegMERGE.r
\name{RvegMerge}
\alias{RvegMerge}
\title{RvegMerge}
\usage{
RvegMerge(x, y, save = "export_merge", head = T)
}
\arguments{
\item{x}{name of first database}

\item{y}{name of second database}

\item{save}{name of exported databes}

\item{head}{logical value if want to merge header}
}
\value{
export two csv files, one for releve and one for header
}
\description{
Merge two Rveg databases
}
