% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RvegMERGE.r
\name{RvegMerge}
\alias{RvegMerge}
\title{RvegMerge}
\usage{
RvegMerge(x, y, save = "export_merge", head = TRUE)
}
\arguments{
\item{x}{name of first database}

\item{y}{name of second database}

\item{save}{name of exported databes}

\item{head}{logical value if want to merge header}
}
\value{
export two csv files, one for releve and one for header
}
\description{
Merge two Rveg databases
}
\examples{
## NOT RUN
if (interactive()) {
  RvegMerge(x = paste0(
    path.package("Rveg"),
    "/extdata/example_db"
  ), y = paste0(
    path.package("Rveg"),
    "/extdata/example_db"
  ))
  read.csv("export_mergeREL.csv", row.names = 1)
}

}
