% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RvegToJuice.R
\name{RvegToJuice}
\alias{RvegToJuice}
\title{RvegToJuice}
\usage{
RvegToJuice(Data, checklist = "default", export = "export")
}
\arguments{
\item{Data}{name of your Rveg database}

\item{checklist}{path to your custom species checklist}

\item{export}{name of your exported csv file}
}
\value{
csv file which is readible by Juice
}
\description{
Export Rveg database to Juice software compatible format
}
