% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RvegToJuice.R
\name{tvToRveg}
\alias{tvToRveg}
\title{TvToRveg}
\usage{
tvToRveg(tv, export = "export", checklist = "default")
}
\arguments{
\item{tv}{path to Turboveg csv export}

\item{export}{name of your exported database}
}
\value{
csv file
}
\description{
Export Turboveg csv file to Rveg database compatible format
}
